updated Controllers

+++ src/main/java/com/one/internship/controller/NotesController.java
@@ -3,11 +3,13 @@ package com.one.internship.controller;
 import com.one.internship.entity.Category;
 import com.one.internship.entity.Note;
 import com.one.internship.entity.User;
+import com.one.internship.model.MessageResponse;
 import com.one.internship.model.NoteInfo;
 import com.one.internship.repository.CategoryRepository;
 import com.one.internship.repository.NoteRepository;
 import com.one.internship.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
@@ -27,10 +29,12 @@ public class NotesController {
     private CategoryRepository categoryRepository;
 
     @GetMapping("/notes")
-    public List<NoteInfo> getNotes(@RequestParam Integer categoryId, @RequestParam String noteContains, Principal principal) {
+    public List<NoteInfo> getNotes(@RequestParam(required = false) Integer categoryId,
+                                   @RequestParam(required = false) String noteContains,
+                                   Principal principal) {
         List<NoteInfo> noteInfos = new ArrayList<>();
         List<Note> noteList = noteRepository.findAll();
-        for (int i = 0; i < noteList.size(); i++){
+        for (int i = 0; i < noteList.size(); i++) {
             NoteInfo noteInfo = new NoteInfo();
             Note n = noteList.get(i);
             noteInfo.setId(n.getNoteId());
@@ -43,19 +47,32 @@ public class NotesController {
     }
 
     @PostMapping("/notes")
-    public void addNotes(@RequestBody NoteInfo req) {
+    public ResponseEntity<MessageResponse> addNotes(@RequestBody NoteInfo req, Principal principal) {
         Note newNote = new Note();
         newNote.setNote(req.getNote());
-        User u = userRepository.findById(req.getOwnerId()).get();
+        User u = userRepository.findByUsername(principal.getName()).get();
         newNote.setOwner(u);
+
         Category c = categoryRepository.findById(req.getCategoryId()).get();
+        if (!c.getOwner().getId().equals(u.getId())) {
+            return ResponseEntity.ok(new MessageResponse("This is not your category"));
+        }
+
         newNote.setCategory(c);
         noteRepository.save(newNote);
+        return ResponseEntity.ok(new MessageResponse("Note added"));
+
     }
 
     @DeleteMapping("/notes/{id}")
-    public void deleteNotes(@PathVariable("id") Integer noteId) {
+    public ResponseEntity<MessageResponse> deleteNotes(@PathVariable("id") Integer noteId, Principal principal) {
+        Note n = noteRepository.findById(noteId).get();
+        User u = userRepository.findByUsername(principal.getName()).get();
+        if (!n.getOwner().getId().equals(u.getId())){
+            return ResponseEntity.ok(new MessageResponse("This is not your note"));
+        }
         noteRepository.deleteById(noteId);
+        return ResponseEntity.ok(new MessageResponse("Note deleted"));
     }
 
 }
